perf: vector store code
