+++ packages/global/common/vectorStore/constants.ts
@@ -1,0 +1,1 @@
+export const PgDatasetTableName = 'modeldata';
