--- packages/service/common/pg/type.d.ts
@@ -1,5 +1,0 @@
-import type { Pool } from 'pg';
-
-declare global {
-  var pgClient: Pool | null;
-}
