+++ projects/app/src/pages/api/core/dataset/searchTest.ts
@@ -6,7 +6,7 @@ import { connectToDatabase } from '@/service/mongo';
 import { authDataset } from '@fastgpt/service/support/permission/auth/dataset';
 import { authTeamBalance } from '@/service/support/permission/auth/bill';
 import { pushGenerateVectorBill } from '@/service/support/wallet/bill/push';
-import { searchDatasetData } from '@/service/core/dataset/data/pg';
+import { searchDatasetData } from '@/service/core/dataset/data/controller';
 import { updateApiKeyUsage } from '@fastgpt/service/support/openapi/tools';
 import { BillSourceEnum } from '@fastgpt/global/support/wallet/bill/constants';
 import { searchQueryExtension } from '@fastgpt/service/core/ai/functions/queryExtension';
@@ -40,7 +40,7 @@ export default withNextCors(async function handler(req: NextApiRequest, res: Nex
     //   model: global.chatModels[0].model
     // });
 
-    const { searchRes, tokenLen } = await searchDatasetData({
+    const { searchRes, tokens } = await searchDatasetData({
       rawQuery: text,
       queries: [text],
       model: dataset.vectorModel,
@@ -53,7 +53,7 @@ export default withNextCors(async function handler(req: NextApiRequest, res: Nex
     const { total } = pushGenerateVectorBill({
       teamId,
       tmbId,
-      tokenLen: tokenLen,
+      tokens,
       model: dataset.vectorModel,
       source: apikey ? BillSourceEnum.api : BillSourceEnum.fastgpt
     });
