+++ projects/app/src/pages/api/v1/embeddings.ts
@@ -5,7 +5,7 @@ import { withNextCors } from '@fastgpt/service/common/middle/cors';
 import { pushGenerateVectorBill } from '@/service/support/wallet/bill/push';
 import { connectToDatabase } from '@/service/mongo';
 import { authTeamBalance } from '@/service/support/permission/auth/bill';
-import { getVectorsByText, GetVectorProps } from '@/service/core/ai/vector';
+import { getVectorsByText, GetVectorProps } from '@fastgpt/service/core/ai/embedding';
 import { updateApiKeyUsage } from '@fastgpt/service/support/openapi/tools';
 import { getBillSourceByAuthType } from '@fastgpt/global/support/wallet/bill/tools';
 
@@ -30,7 +30,7 @@ export default withNextCors(async function handler(req: NextApiRequest, res: Nex
 
     await authTeamBalance(teamId);
 
-    const { tokenLen, vectors } = await getVectorsByText({ input, model });
+    const { tokens, vectors } = await getVectorsByText({ input, model });
 
     jsonRes(res, {
       data: {
@@ -42,8 +42,8 @@ export default withNextCors(async function handler(req: NextApiRequest, res: Nex
         })),
         model,
         usage: {
-          prompt_tokens: tokenLen,
-          total_tokens: tokenLen
+          prompt_tokens: tokens,
+          total_tokens: tokens
         }
       }
     });
@@ -51,7 +51,7 @@ export default withNextCors(async function handler(req: NextApiRequest, res: Nex
     const { total } = pushGenerateVectorBill({
       teamId,
       tmbId,
-      tokenLen,
+      tokens,
       model,
       billId,
       source: getBillSourceByAuthType({ authType })
