+++ projects/app/src/service/core/dataset/utils.ts
@@ -1,6 +1,4 @@
 import { MongoDatasetData } from '@fastgpt/service/core/dataset/data/schema';
-import { cut } from '@node-rs/jieba';
-import { stopWords } from '@fastgpt/global/common/string/jieba';
 
 /**
  * Same value judgment
@@ -23,15 +21,4 @@ export async function hasSameValue({
   if (count > 0) {
     return Promise.reject('已经存在完全一致的数据');
   }
-}
-
-export function jiebaSplit({ text }: { text: string }) {
-  const tokens = cut(text, true);
-
-  return (
-    tokens
-      .map((item) => item.replace(/[^\u4e00-\u9fa5a-zA-Z0-9\s]/g, '').trim())
-      .filter((item) => item && !stopWords.has(item))
-      .join(' ') || ''
-  );
 }
