+++ projects/app/src/service/mongo.ts
@@ -1,11 +1,11 @@
 import { startQueue } from './utils/tools';
 import { PRICE_SCALE } from '@fastgpt/global/support/wallet/bill/constants';
-import { initPg } from '@fastgpt/service/common/pg';
 import { MongoUser } from '@fastgpt/service/support/user/schema';
 import { connectMongo } from '@fastgpt/service/common/mongo/init';
 import { hashStr } from '@fastgpt/global/common/string/tools';
 import { createDefaultTeam } from '@fastgpt/service/support/user/team/controller';
 import { exit } from 'process';
+import { initVectorStore } from '@fastgpt/service/common/vectorStore/controller';
 
 /**
  * connect MongoDB and init data
@@ -14,7 +14,7 @@ export function connectToDatabase(): Promise<void> {
   return connectMongo({
     beforeHook: () => {},
     afterHook: () => {
-      initPg();
+      initVectorStore();
       // start queue
       startQueue();
       return initRootUser();
